requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
azure-search-documents==11.6.0b12
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==4.2.21
djangorestframework==3.16.0
dotenv==0.9.9
exceptiongroup==1.3.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.43
lxml==5.4.0
msrest==0.7.1
oauthlib==3.2.2
openai==1.82.1
orjson==3.10.18
packaging==24.2
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
XlsxWriter==3.2.3
zstandard==0.23.0
